requirements.txt
asgiref==3.10.0
black==25.11.0
cfgv==3.4.0
click==8.1.8
distlib==0.4.0
Django==4.2.26
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==37.12.0
filelock==3.19.1
flake8==7.3.0
identify==2.6.15
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
isort==6.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-factoryboy==2.8.1
pytokens==0.3.0
PyYAML==6.0.3
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.3
virtualenv==20.35.4
zipp==3.23.0
